+++ Example/HCaptcha_Tests/Core/HCaptchaResult__Tests.swift
@@ -13,7 +13,8 @@ import XCTest
 class HCaptchaResult__Tests: XCTestCase {
     func test__Get_Token() {
         let token = UUID().uuidString
-        let result = HCaptchaResult(token: token)
+        let manager = HCaptchaWebViewManager()
+        let result = HCaptchaResult(manager, token: token)
 
         do {
             let value = try result.dematerialize()
@@ -26,7 +27,8 @@ class HCaptchaResult__Tests: XCTestCase {
 
     func test__Get_Token__Error() {
         let error = HCaptchaError.random()
-        let result = HCaptchaResult(error: error)
+        let manager = HCaptchaWebViewManager()
+        let result = HCaptchaResult(manager, error: error)
 
         do {
             _ = try result.dematerialize()